+++ typescript/packages/plugins/erc721/src/tools.ts
@@ -1,6 +1,5 @@
-import { allowance, approve, balanceOf, totalSupply, transfer, transferFrom } from "./methods";
+import { approve, balanceOf, totalSupply, transfer, transferFrom } from "./methods";
 import {
-    allowanceParametersSchema,
     approveParametersSchema,
     getBalanceParametersSchema,
     totalSupplyParametersSchema,
@@ -20,7 +19,7 @@ export function getTools(walletClient: EVMWalletClient, tokenList: ChainSpecific
         const balanceTool = createTool(
             {
                 name: `get_${token.symbol}_balance`,
-                description: `Get the balance of ${token.symbol}`,
+                description: `This {{tool}} gets the number of NFTs owned for ${token.symbol}`,
                 parameters: getBalanceParametersSchema,
             },
             (parameters: z.infer<typeof getBalanceParametersSchema>) => balanceOf(walletClient, token, parameters),
@@ -29,7 +28,7 @@ export function getTools(walletClient: EVMWalletClient, tokenList: ChainSpecific
         const transferTool = createTool(
             {
                 name: `transfer_${token.symbol}`,
-                description: `Transfer ${token.symbol} to the specified address`,
+                description: `This {{tool}} transfers a ${token.symbol} NFT to the specified address`,
                 parameters: transferParametersSchema,
             },
             (parameters: z.infer<typeof transferParametersSchema>) => transfer(walletClient, token, parameters),
@@ -38,25 +37,16 @@ export function getTools(walletClient: EVMWalletClient, tokenList: ChainSpecific
         const totalSupplyTool = createTool(
             {
                 name: `get_${token.symbol}_total_supply`,
-                description: `Get the total supply of ${token.symbol}`,
+                description: `This {{tool}} gets the total supply of ${token.symbol} NFTs`,
                 parameters: totalSupplyParametersSchema,
             },
             (parameters: z.infer<typeof totalSupplyParametersSchema>) => totalSupply(walletClient, token),
         );
 
-        const allowanceTool = createTool(
-            {
-                name: `get_${token.symbol}_allowance`,
-                description: `Get the allowance of ${token.symbol}`,
-                parameters: allowanceParametersSchema,
-            },
-            (parameters: z.infer<typeof allowanceParametersSchema>) => allowance(walletClient, token, parameters),
-        );
-
         const approveTool = createTool(
             {
                 name: `approve_${token.symbol}`,
-                description: `Approve the allowance of ${token.symbol}`,
+                description: `This {{tool}} approves an address to transfer a specific ${token.symbol} NFT`,
                 parameters: approveParametersSchema,
             },
             (parameters: z.infer<typeof approveParametersSchema>) => approve(walletClient, token, parameters),
@@ -65,13 +55,13 @@ export function getTools(walletClient: EVMWalletClient, tokenList: ChainSpecific
         const transferFromTool = createTool(
             {
                 name: `transfer_${token.symbol}_from`,
-                description: `Transfer ${token.symbol} from the specified address`,
+                description: `This {{tool}} transfers a ${token.symbol} NFT from one address to another`,
                 parameters: transferFromParametersSchema,
             },
             (parameters: z.infer<typeof transferFromParametersSchema>) => transferFrom(walletClient, token, parameters),
         );
 
-        tools.push(balanceTool, transferTool, totalSupplyTool, allowanceTool, approveTool, transferFromTool);
+        tools.push(balanceTool, transferTool, totalSupplyTool, approveTool, transferFromTool);
     }
 
     return tools;