+++ typescript/examples/langchain/lit/CHANGELOG.md
@@ -1,5 +1,34 @@
 # goat-examples-langchain-lit
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [65ee1e8]
+  - @goat-sdk/wallet-lit@0.2.5
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [c565119]
+  - @goat-sdk/plugin-erc20@0.2.2
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [de8533d]
+  - @goat-sdk/wallet-solana@0.2.4
+  - @goat-sdk/wallet-lit@0.2.4
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [f10e6d1]
+  - @goat-sdk/adapter-langchain@0.2.1
+
 ## 0.1.3
 
 ### Patch Changes